+++ pyhap/__init__.py
@@ -1,10 +1,10 @@
 import os
 
-_ROOT = os.path.abspath(os.path.dirname(__file__))
-_RESOURCE_DIR = os.path.join(_ROOT, "resources")
+ROOT = os.path.abspath(os.path.dirname(__file__))
+RESOURCE_DIR = os.path.join(ROOT, "resources")
 
-CHARACTERISTICS_FILE = os.path.join(_RESOURCE_DIR, "characteristics.json")
-SERVICES_FILE = os.path.join(_RESOURCE_DIR, "services.json")
+CHARACTERISTICS_FILE = os.path.join(RESOURCE_DIR, "characteristics.json")
+SERVICES_FILE = os.path.join(RESOURCE_DIR, "services.json")
 
 
 # Flag if QR Code dependencies are installed.