+++ CHANGELOG.md
@@ -16,6 +16,27 @@ Sections
 ### Developers
 -->
 
+## [2.3.0] - 2018-10-25
+
+### Added
+- Added support for the camera accessory. [#161](https://github.com/ikalchev/HAP-python/pull/161)
+- Added a demo script that starts four fake accessories.
+- Added `NeoPixelsLightStrip` accessory. [#144](https://github.com/ikalchev/HAP-python/pull/144)
+- Added new Accessory categories. [Commit](https://github.com/ikalchev/HAP-python/commit/dbf1d5d8fea814af52098f3a2f3cd5a47cd889c1)
+
+### Changed
+- Updated the README with information on how to setup a camera accessory.
+- Spelling fix - executor (accessory_driver). [#159](https://github.com/ikalchev/HAP-python/pull/159)
+- Char.client_update_value now ignores call if value is already set. This could happen during automations. [#162](https://github.com/ikalchev/HAP-python/pull/162)
+
+### Fixed
+- Updated README. [#138](https://github.com/ikalchev/HAP-python/pull/138)
+- Accessory return codes for characteristics. [#143](https://github.com/ikalchev/HAP-python/pull/143)
+- Deprecation notice in `accessory.py/config_changed`. [#150](https://github.com/ikalchev/HAP-python/pull/150)
+- Add end_response in set_characteristics. [#153](https://github.com/ikalchev/HAP-python/pull/153)
+
+
+
 ## [2.2.2] - 2018-05-29
 
 ### Fixed